ai-service/pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelforge-ai-service"
version = "0.1.0"
description = "AI service for PixelForge image operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "pydantic-settings==2.10.1",
  "python-json-logger==3.3.0",
  "rembg==2.0.75",
  "pillow==12.1.0",
  "python-multipart==0.0.20"
]

[project.optional-dependencies]
dev = [
  "pytest==8.4.2",
  "httpx==0.28.1",
  "pytest-asyncio==1.1.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
